backend/pyproject.toml
[tool.poetry]
name = "poolstats"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
flask = "^1.1.1"
schedule = "^0.6.0"
BeautifulSoup4 = "^4.8.2"
requests = "^2.23.0"
flask_cors = "^3.0.8"
python-dateutil = "^2.8.1"
pytz = "^2019.3"
peewee = "^3.13.1"
gunicorn = "^20.0.4"
flask_script = "^2.0.6"

[tool.poetry.dev-dependencies]
invoke = "^1.4.1"
patchwork = "^1.0.1"
fabric = "^2.5.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
